requirements.txt
requests
flask
imageio==1.5
numpy
Pillow==3.3.1
gunicorn==19.8.1
